Split out the function for getting a timetable indexed by time and day into a new utils module
